\name{cssTools-package}
\alias{cssTools-package}
\alias{cssTools}
\docType{package}
\title{
\packageTitle{cssTools}
}
\description{
Formalized by Krackhardt (1987), Cognitive Social Structure (CSS) network studies collect relational data on respondents direct ties 
and their cognition of ties among all other individuals in the network. This package provides a collection of tools for estimating a 
network from a random sample of CSS slices. The package also contains functions for evaluating a CSS in terms of various 
error types observed in each slice.
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{cssTools}
\packageIndices{cssTools}
}
\author{
\packageAuthor{cssTools}
Maintainer: \packageMaintainer{cssTools}
}
\references{
Krackhardt, D. (1987). Cognitive social structures. Social Networks 9, 109-134. \url{http://dx.doi.org/10.1016/0378-8733(87)90009-8}

D. Yenigun, G. Ertan, M.D. Siciliano (2016). Omission and commission errors in 
network cognition and estimation using ROC curve. arXiv:1606.03245 [stat.CO] \url{https://arxiv.org/abs/1606.03245}
}
\keyword{ package }
\seealso{
\code{\link{atm}}, \code{\link{cssTools2sna}}, \code{\link{ftm}}, 
\code{\link{highTechManagers}}, \code{\link{rtm}}, \code{\link{rtmPlot}}, 
\code{\link{s14}}, \code{\link{sliceQuality}}, \code{\link{sna2cssTools}}
}

